chatbot/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
langchain-text-splitters==0.3.4
langchain-core==0.3.29
langchain-huggingface>=0.1.0
chromadb==0.5.23
sentence-transformers==3.3.1
reportlab